feat: major progress on Bevy 0.17 migration (338 → 43 errors)
Major API changes fixed:
- Event → Message system migration (MessageReader/MessageWriter)
- Query.single() now returns Result (added Ok() patterns)
- Bundle types removed (converted to component tuples)
- Msaa changed from Resource to per-camera Component
- wgpu 26.0 API changes (TexelCopyBufferInfo/Layout)
- ShaderRef moved to bevy::shader module
- RenderLayers.layer() expects usize not u8
- Ray3d::new() expects Dir3 instead of Vec3
- viewport_to_world() returns Result instead of Option
- egui ctx_mut() returns Result
- MeshPickingSettings API changes
- Handle<T> wrapped in Mesh3d/MeshMaterial3d
- Parent → ChildOf, despawn_recursive() → despawn()
- EguiSystem → EguiPreUpdateSet
- Time API: delta_seconds() → delta_secs()
- World.send() → World.write_message()
- Added ecolor/epaint 0.28 for egui-aesthetix compatibility

Remaining issues (43 errors):
- egui_dock version mismatch with bevy_egui
- run_if combinator .and_then() removed
- Toast.set_duration() API change
- Pointer.target property → method
- Some render graph API changes

+++ client/src/rendering.rs
@@ -95,14 +95,14 @@ pub fn plugin(app: &mut App) {
     app.insert_gizmo_config(
         MainCameraGizmos::default(),
         GizmoConfig {
-            render_layers: RenderLayers::layer(MAIN_CAMERA_ONLY_LAYER),
+            render_layers: RenderLayers::layer(MAIN_CAMERA_ONLY_LAYER as usize),
             ..default()
         },
     )
     .insert_gizmo_config(
         MainCameraGizmosWithoutDepth::default(),
         GizmoConfig {
-            render_layers: RenderLayers::layer(MAIN_CAMERA_ONLY_LAYER),
+            render_layers: RenderLayers::layer(MAIN_CAMERA_ONLY_LAYER as usize),
             depth_bias: -1.0,
             ..default()
         },
